requirements.txt
absl-py==1.4.0
astor==0.8.1
black==23.3.0
click==8.1.3
cycler==0.11.0
fonttools==4.38.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
importlib-metadata==6.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
mypy-extensions==1.0.0
numpy==1.19.5
opt-einsum==3.3.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
protobuf==3.20.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==2.3.0
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.5.0
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
